WIP function.

+++ src/Script_tbl_map_8002BB44.h
@@ -1,13 +1,14 @@
 #ifndef SCRIPT_TBL_MAP_H
 #define SCRIPT_TBL_MAP_H
 
+#include <util/idaTypes.h>
 
 typedef struct Unk_800B4D98
 {
     short field_0;                // 800b4d98
     short field_2_DiffcultyLevel; // 800b4d9a
-    int Flags;                    // 800b4d9c
-    short StartingCdId;           // 800b4d9c
+    WORD Flags;                   // 800b4d9c
+    short StartingCdId;           // 800b4d9e
     short LastRand;               // 800b4da0
 } Unk_800B4D98;
